Making more components use DSA stuff, better integrating DSA shader stuff, WIP shader compiler, cleanup elsewhere (so much cleanup still to do)...

+++ DiamondDogs/scenes/StarDemo.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "StarDemo.h"
 
-demo::StarDemo::StarDemo() : Context(SCR_WIDTH, SCR_HEIGHT), demoSkybox(skyboxTextures), demoStar(4, 300.0f, 1000, Projection) {
+demo::StarDemo::StarDemo() : Context((GLfloat)SCR_WIDTH, (GLfloat)SCR_HEIGHT), demoSkybox(skyboxTextures), demoStar(4, 300.0f, 1000, Projection) {
 	demoSkybox.BuildRenderData(Projection);
 }
